chore(ClipboardButton): extract `<FakeButton/>`
