+++ src/components/ClipboardButton/ClipboardButton.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
-import {block} from '../utils/cn';
 import {CopyToClipboard} from '../CopyToClipboard';
-import {ClipboardIcon} from '../ClipboardIcon';
 import {QAProps} from '../types';
-import './ClipboardButton.scss';
+import {FakeButton} from './FakeButton';
 
 export interface ClipboardButtonProps extends QAProps {
     /** Текст для копирования */
@@ -14,21 +12,11 @@ export interface ClipboardButtonProps extends QAProps {
     className?: string;
 }
 
-const b = block('clipboard-button');
-const DEFAULT_ICON_SIZE = 24;
-
-export function ClipboardButton({
-    text,
-    size = DEFAULT_ICON_SIZE,
-    className,
-    qa,
-}: ClipboardButtonProps) {
+export function ClipboardButton({text, size, className, qa}: ClipboardButtonProps) {
     return (
         <CopyToClipboard text={text} timeout={1000}>
             {(status) => (
-                <span className={b(null, className)} data-qa={qa}>
-                    <ClipboardIcon status={status} size={size} className={b('icon')} />
-                </span>
+                <FakeButton className={className} data-qa={qa} status={status} size={size} />
             )}
         </CopyToClipboard>
     );
